154z90

+++ src/drivers/Adafruit_SSD1681.cpp
@@ -41,7 +41,7 @@ Adafruit_SSD1681::Adafruit_SSD1681(int width, int height, int8_t SID,
     height += 8 - (height % 8);
   }
 
-  buffer1_size = width * height / 8;
+  buffer1_size = ((uint32_t)width * (uint32_t)height) / 8;
   buffer2_size = buffer1_size;
 
   if (SRCS >= 0) {
@@ -79,7 +79,7 @@ Adafruit_SSD1681::Adafruit_SSD1681(int width, int height, int8_t DC, int8_t RST,
     height += 8 - (height % 8);
   }
 
-  buffer1_size = width * height / 8;
+  buffer1_size = ((uint32_t)width * (uint32_t)height) / 8;
   buffer2_size = buffer1_size;
 
   if (SRCS >= 0) {
